backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.9.2
asttokens==3.0.0
attrs==25.3.0
beautifulsoup4==4.14.2
billiard==4.2.2
boto3==1.40.45
botocore==1.40.45
browserforge==1.2.3
camoufox==0.4.11
celery==5.5.3
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
contextlib2==21.6.0
cron_descriptor==2.0.6
cryptography==46.0.2
cssselect==1.3.0
curl_cffi==0.13.0
decorator==5.2.1
Django==5.2.7
django-celery-beat==2.8.1
django-cors-headers==4.9.0
django-ninja==1.4.3
django-ninja-extra==0.30.1
django-ninja-jwt==5.3.9
django-storages==1.14.6
django-timezone-field==7.1
django_celery_results==2.6.0
executing==2.2.1
filelock==3.19.1
frozenlist==1.7.0
geoip2==5.1.0
greenlet==3.2.4
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
injector==0.22.0
ipython==9.6.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.5.4
language-tags==1.2.0
lxml==6.0.2
markdownify==1.2.0
matplotlib-inline==0.1.7
maxminddb==2.8.2
mcp==1.15.0
msgspec==0.19.0
multidict==6.6.4
numpy==2.3.3
orjson==3.11.3
packaging==25.0
parso==0.8.5
patchright==1.55.2
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.4.0
playwright==1.55.0
prompt_toolkit==3.0.52
propcache==0.3.2
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.23
pydantic==2.11.9
pydantic-settings==2.11.0
pydantic_core==2.33.2
pyee==13.0.0
Pygments==2.19.2
PyJWT==2.10.1
PySocks==1.7.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
redis==6.4.0
referencing==0.36.2
requests==2.32.5
requests-file==2.1.0
rpds-py==0.27.1
s3transfer==0.14.0
scrapling==0.3.6
screeninfo==0.8.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
sqlparse==0.5.3
sse-starlette==3.0.2
stack-data==0.6.3
starlette==0.48.0
tldextract==5.3.0
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
urllib3==2.5.0
uvicorn==0.37.0
vine==5.1.0
wcwidth==0.2.14
yarl==1.20.1
